Fix audio desync correction
Discarding samples from our audio buffer doesn't discard samples from the
audio device interface responsible for actually playing sound. This means
that if the time difference is large enough, we can get stuck attempting
to discard samples from our buffer because all of that time difference is
in the device buffer, which we can't directly manipulate.

Instead, we stop the player to force it to flush samples, and then restart
it after discarding the samples we need to get rid of.